moved history tests from collection to history reducer
