+++ packages/opds-browser/src/reducers/__tests__/collection-test.ts
@@ -1,6 +1,8 @@
 jest.dontMock("../collection");
-jest.dontMock("../history");
 jest.dontMock("../../actions");
+jest.setMock("../history", {
+  default: (state, action) => state.history
+});
 
 import reducer from "../collection";
 import DataFetcher from "../../DataFetcher";
@@ -127,260 +129,10 @@ describe("collection reducer", () => {
     let newState = Object.assign({}, currentState, {
       url: "some other url",
       data: data,
-      isFetching: false,
-      history: [{
-        id: "id",
-        text: "title",
-        url: "url"
-      }]
-    });
-
-    expect(reducer(currentState, action)).toEqual(newState);
-  });
-
-  it("should handle LOAD_COLLECTION if new parent url is equal to previous colllection url", () => {
-    let stateWithHistory = Object.assign({}, currentState, {
-      history: [{
-        id: null,
-        url: "root url",
-        text: "root title"
-      }, {
-        id: "test id",
-        url: "test url",
-        text: "test title"
-      }]
-    });
-    let data = {
-      id: "some id",
-      url: "some url",
-      title: "some title",
-      lanes: [],
-      books: [],
-      links: [],
-      catalogRootLink: stateWithHistory.data.catalogRootLink,
-      parentLink: {
-        url: stateWithHistory.data.url,
-        text: stateWithHistory.data.title
-      }
-    };
-    let action = actions.loadCollection(data, "some url");
-    let newState = Object.assign({}, currentState, {
-      url: "some url",
-      data: data,
-      isFetching: false,
-      history: stateWithHistory.history.concat({
-        id: stateWithHistory.data.id,
-        text: stateWithHistory.data.title,
-        url: stateWithHistory.data.url
-      })
-    });
-
-    expect(reducer(stateWithHistory, action)).toEqual(newState);
-  });
-
-  it("shouldn't change history on LOAD_COLLECTION with same id", () => {
-    let data = {
-      id: "id",
-      url: "some url",
-      title: "some title",
-      lanes: [],
-      books: [],
-      links: []
-    };
-    let action = actions.loadCollection(data, "some other url");
-    let newState = Object.assign({}, currentState, {
-      url: "some other url",
-      data: data,
       isFetching: false
     });
 
     expect(reducer(currentState, action)).toEqual(newState);
-  });
-
-  it("should clear history on LOAD_COLLECTION with the old catalog root", () => {
-    let stateWithHistory = Object.assign({}, currentState, {
-      history: [{
-        id: "test id",
-        url: "test url",
-        text: "test title"
-      }]
-    });
-    let data = {
-      id: "some id",
-      url: "root url",
-      title: "root title",
-      lanes: [],
-      books: [],
-      links: []
-    };
-    let action = actions.loadCollection(data, "root url");
-    let newState = Object.assign({}, currentState, {
-      url: "root url",
-      data: data,
-      isFetching: false,
-      history: []
-    });
-
-    expect(reducer(stateWithHistory, action)).toEqual(newState);
-  });
-
-  it("should clear history on LOAD_COLLECTION with a new catalog", () => {
-    let stateWithHistory = Object.assign({}, currentState, {
-      history: [{
-        id: "test id",
-        url: "test url",
-        title: "test title"
-      }]
-    });
-    let data = {
-      id: "some id",
-      url: "some url",
-      title: "some title",
-      catalogRootLink: {
-        url: "new root url",
-        text: "new root title"
-      },
-      lanes: [],
-      books: [],
-      links: []
-    };
-    let action = actions.loadCollection(data, "some url");
-    let newState = Object.assign({}, stateWithHistory, {
-      url: "some url",
-      data: data,
-      isFetching: false,
-      history: [{
-        id: null,
-        url: "new root url",
-        text: "new root title"
-      }]
-    });
-
-    expect(reducer(stateWithHistory, action)).toEqual(newState);
-  });
-
-  it("should set history to root and parent on LOAD_COLLECTION if it's not there", () => {
-    let data = {
-      id: "some id",
-      url: "some url",
-      title: "some title",
-      lanes: [],
-      books: [],
-      links: [],
-      catalogRootLink: {
-        url: "root url",
-        text: "root title"
-      },
-      parentLink: {
-        url: "parent url",
-        text: "parent title"
-      }
-    };
-    let action = actions.loadCollection(data, "some url");
-    let newState = Object.assign({}, currentState, {
-      url: "some url",
-      data: data,
-      isFetching: false,
-      history: [{
-        text: "root title",
-        url: "root url",
-        id: null
-      }, {
-        text: "parent title",
-        url: "parent url",
-        id: null
-      }]
-    });
-
-    expect(reducer(currentState, action)).toEqual(newState);
-  });
-
-  it("should set history to root and parent on LOAD_COLLECTION if parent url is different than last history url", () => {
-    let oldState = Object.assign({}, currentState, {
-      history: [
-        {
-          id: null,
-          text: "old root title",
-          url: "old root url"
-        }, {
-          id: "some other id",
-          url: "some other url",
-          text: "some other title"
-        }
-      ]
-    });
-    let data = {
-      id: "some id",
-      url: "some url",
-      title: "some title",
-      lanes: [],
-      books: [],
-      links: [],
-      catalogRootLink: {
-        url: "root url",
-        text: "root title"
-      },
-      parentLink: {
-        url: "parent url",
-        text: "parent title"
-      }
-    };
-    let action = actions.loadCollection(data, "some url");
-    let newState = Object.assign({}, oldState, {
-      url: "some url",
-      data: data,
-      isFetching: false,
-      history: [{
-        id: null,
-        text: "root title",
-        url: "root url"
-      }, {
-        id: null,
-        text: "parent title",
-        url: "parent url"
-      }]
-    });
-
-    expect(reducer(oldState, action)).toEqual(newState);
-  });
-
-  it("should remove history up to loaded url on LOAD_COLLECTION with url in history", () => {
-    let stateWithHistory = Object.assign({}, currentState, {
-      history: [{
-        id: "first id",
-        url: "first url",
-        title: "first title"
-      }, {
-        id: "test id",
-        url: "test url",
-        title: "test title"
-      }, {
-        id: "other id",
-        url: "other url",
-        title: "other title"
-      }]
-    });
-    let data = {
-      id: "some id",
-      url: "test url",
-      title: "some title",
-      lanes: [],
-      books: [],
-      links: []
-    };
-    let action = actions.loadCollection(data, "test url");
-    let newState = Object.assign({}, currentState, {
-      url: "test url",
-      data: data,
-      isFetching: false,
-      history: [{
-        id: "first id",
-        url: "first url",
-        title: "first title"
-      }]
-    });
-
-    expect(reducer(stateWithHistory, action)).toEqual(newState);
   });
 
   it("should handle LOAD_COLLECTION after an error", () => {
